[Refactor] 배열 변환 함수 추출
